Revert "Dev hector"
