+++ package.json
@@ -19,11 +19,9 @@
     "@angular/platform-browser": "~10.0.0",
     "@angular/platform-browser-dynamic": "~10.0.0",
     "@angular/router": "~10.0.0",
-    "@sweetalert2/themes": "^5.0.15",
     "chart.js": "^2.9.3",
     "ng2-charts": "^2.3.2",
     "rxjs": "~6.5.4",
-    "sweetalert2": "^11.6.16",
     "tslib": "^2.0.0",
     "zone.js": "~0.10.2"
   },
