+++ src/app/auth/login/login.component.html
@@ -4,57 +4,39 @@
 <section id="wrapper" class="login-register login-sidebar" style="background-image:url(../assets/images/background/login-register.jpg);">
     <div class="login-box card">
         <div class="card-body">
-            <form class="form-horizontal form-material"
-                  (submit)="login()"
-                  [formGroup]="loginForm"
-                  id="loginform">
+            <form class="form-horizontal form-material" id="loginform" action="index.html">
                 <a href="javascript:void(0)" class="text-center db"><img src="../assets/images/logo-icon.png" alt="Home" /><br/><img src="../assets/images/logo-text.png" alt="Home" /></a>
                 <div class="form-group m-t-40">
                     <div class="col-xs-12">
-                        <input class="form-control"
-                               type="email"
-                               formControlName="email"
-                               placeholder="Email">
+                        <input class="form-control" type="text" required="" placeholder="Username">
                     </div>
                 </div>
                 <div class="form-group">
                     <div class="col-xs-12">
-                        <input class="form-control"
-                               type="password"
-                               formControlName="password"
-                               placeholder="Contraseña">
+                        <input class="form-control" type="password" required="" placeholder="Password">
                     </div>
                 </div>
                 <div class="form-group row">
                     <div class="col-md-12">
                         <div class="checkbox checkbox-primary pull-left p-t-0">
-                            <input id="checkbox-signup" type="checkbox" class="filled-in chk-col-light-blue" formControlName="remember">
+                            <input id="checkbox-signup" type="checkbox" class="filled-in chk-col-light-blue">
                             <label for="checkbox-signup"> Remember me </label>
                         </div>
                         <a href="javascript:void(0)" id="to-recover" class="text-dark pull-right"><i class="fa fa-lock m-r-5"></i> Forgot pwd?</a> </div>
                 </div>
-
                 <div class="form-group text-center m-t-20">
                     <div class="col-xs-12">
                         <button class="btn btn-info btn-lg btn-block text-uppercase btn-rounded" type="submit">Log In</button>
                     </div>
                 </div>
-
                 <div class="row">
                     <div class="col-xs-12 col-sm-12 col-md-12 m-t-10 text-center">
-                        <div class="social">
-                          <div #googleBtn id="buttonDiv"></div>
-                          <!-- <a href="javascript:void(0)"
-                             class="btn btn-googleplus"
-                             data-toggle="tooltip"
-                             title="Login with Google">
-                          <i aria-hidden="true" class="fa fa-google-plus"></i> </a> -->
-                        </div>
+                        <div class="social"><a href="javascript:void(0)" class="btn  btn-facebook" data-toggle="tooltip" title="Login with Facebook"> <i aria-hidden="true" class="fa fa-facebook"></i> </a> <a href="javascript:void(0)" class="btn btn-googleplus" data-toggle="tooltip" title="Login with Google"> <i aria-hidden="true" class="fa fa-google-plus"></i> </a> </div>
                     </div>
                 </div>
                 <div class="form-group m-b-0">
                     <div class="col-sm-12 text-center">
-                        Don't have an account? <a routerLink="/register" class="text-primary m-l-5"><b>Sign Up</b></a>
+                        Don't have an account? <a href="pages-register2.html" class="text-primary m-l-5"><b>Sign Up</b></a>
                     </div>
                 </div>
             </form>
