+++ src/app/auth/register/register.component.html
@@ -4,67 +4,37 @@
 <section id="wrapper" class="login-register login-sidebar" style="background-image:url(../assets/images/background/login-register.jpg);">
     <div class="login-box card">
         <div class="card-body">
-            <form  class="form-horizontal form-material"
-                   id="loginform"
-                   [formGroup]="registerForm"
-                   (ngSubmit)="crearUsuario()"
-                   >
-
+            <form class="form-horizontal form-material" id="loginform" action="index.html">
                 <a href="javascript:void(0)" class="text-center db"><img src="../assets/images/logo-icon.png" alt="Home" /><br/><img src="../assets/images/logo-text.png" alt="Home" /></a>
                 <h3 class="box-title m-t-40 m-b-0">Register Now</h3><small>Create your account and enjoy</small>
                 <div class="form-group m-t-20">
-
                     <div class="col-xs-12">
-                        <input class="form-control"
-                               type="text"
-                               placeholder="Name"
-                               formControlName="nombre">
-                    </div>
-
-                </div>
-                <div class="form-group ">
-                    <div class="col-xs-12">
-                        <input class="form-control"
-                               type="text"
-                               placeholder="Email"
-                               formControlName="email">
+                        <input class="form-control" type="text" required="" placeholder="Name">
                     </div>
                 </div>
                 <div class="form-group ">
                     <div class="col-xs-12">
-                        <input class="form-control"
-                               type="password"
-                               placeholder="Password"
-                               formControlName="password">
+                        <input class="form-control" type="text" required="" placeholder="Email">
+                    </div>
+                </div>
+                <div class="form-group ">
+                    <div class="col-xs-12">
+                        <input class="form-control" type="password" required="" placeholder="Password">
                     </div>
                 </div>
                 <div class="form-group">
                     <div class="col-xs-12">
-                        <input class="form-control"
-                               type="password"
-                               placeholder="Confirm Password"
-                               formControlName="password2">
+                        <input class="form-control" type="password" required="" placeholder="Confirm Password">
                     </div>
                 </div>
                 <div class="form-group row">
                     <div class="col-md-12">
                         <div class="checkbox checkbox-primary p-t-0">
-                            <input id="checkbox-signup" type="checkbox" formControlName="terminos">
+                            <input id="checkbox-signup" type="checkbox">
                             <label for="checkbox-signup"> I agree to all <a href="#">Terms</a></label>
                         </div>
                     </div>
                 </div>
-
-                <div class="form-group row">
-                  <div class="col-md-12 text-danger">
-                    <p *ngIf="camponoValido('nombre')">El nombre es obligatorio</p>
-                    <p *ngIf="camponoValido('email')">El email es obligatorio</p>
-                    <p *ngIf="contrasenasNoValidas()">Las Contraseñas deben ser iguales</p>
-                    <!-- <p *ngIf="contrasenasNoValidas('password2')">La confirmacion del password es obligatorio</p> -->
-                    <p *ngIf="aceptaTerminos()">Debe aceptar los terminos</p>
-                  </div>
-                </div>
-
                 <div class="form-group text-center m-t-20">
                     <div class="col-xs-12">
                         <button class="btn btn-info btn-lg btn-block text-uppercase waves-effect waves-light" type="submit">Sign Up</button>
@@ -72,7 +42,7 @@
                 </div>
                 <div class="form-group m-b-0">
                     <div class="col-sm-12 text-center">
-                        <p>Already have an account? <a routerLink="/login" class="text-info m-l-5"><b>Sign In</b></a></p>
+                        <p>Already have an account? <a href="pages-login-2.html" class="text-info m-l-5"><b>Sign In</b></a></p>
                     </div>
                 </div>
             </form>
