--- src/app/interfaces/login-form.interface.ts
@@ -1,5 +1,0 @@
-export interface LoginForm{
-  email: string;
-  password: string;
-  remember: boolean;
-}
