--- src/app/interfaces/register-form.interface..ts
@@ -1,9 +1,0 @@
-
-
-export interface RegisterForm{
-    nombre: string;
-    email: string;
-    password: string;
-    password2: string;
-    terminos: boolean;
-}
