--- src/app/models/usuarios.model.ts
@@ -1,29 +1,0 @@
-import { environment } from 'src/environments/environment';
-
-const api_url = environment.base_url;
-
-export class Usuario {
-
-  constructor(
-      public nombre: string,
-      public email: string,
-      public password: string,
-      public google?: boolean,
-      public img?: string,
-      public role?: string,
-      public uid?: string,
-    ){}
-
-    get imagenUrl() {
-
-      if( this.img.includes('https') ){
-        return this.img;
-      }
-
-      if( this.img ){
-        return `${api_url}/upload/usuarios/${ this.img }`
-      }else{
-        return `${api_url}/upload/usuarios/no-image`
-      }
-    }
-}
