+++ src/app/pages/pages.component.html
@@ -19,7 +19,7 @@
                 <div class="col-12">
                     <div class="card">
                         <div class="card-body">
-
+                            
                         </div>
                     </div>
                 </div>
@@ -31,7 +31,7 @@
         <!-- footer -->
         <!-- ============================================================== -->
         <footer class="footer">
-            © 2022 Admin Pro by wrappixel.com
+            © 20xx Admin Pro by wrappixel.com
         </footer>
         <!-- ============================================================== -->
         <!-- End footer -->
