+++ src/app/pages/pages.component.ts
@@ -1,7 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { SettingsService } from '../services/settings.service';
-
-declare function customInitFunctions();
 
 @Component({
   selector: 'app-pages',
@@ -11,12 +8,13 @@ declare function customInitFunctions();
 })
 export class PagesComponent implements OnInit {
 
-  constructor(
-    private settingsService: SettingsService
-  ) { }
+  linkTheme: Element = document.querySelector('#theme')
+
+  constructor() { }
 
   ngOnInit(): void {
-    customInitFunctions();
+    const theme = localStorage.getItem('theme') || "./assets/css/colors/default-dark.css";
+    this.linkTheme.setAttribute('href', theme);
   }
