+++ src/app/pages/pages.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms'
+import { FormsModule } from '@angular/forms'
 
 // Modulos
 import { SharedModule } from '../shared/shared.module';
@@ -12,7 +12,6 @@ import { ProgressComponent } from './progress/progress.component';
 import { Grafica1Component } from './grafica1/grafica1.component';
 import { PagesComponent } from './pages.component';
 import { AccountSettingsComponent } from './account-settings/account-settings.component';
-import { PerfilComponent } from './perfil/perfil.component';
 
 
 
@@ -24,7 +23,6 @@ import { PerfilComponent } from './perfil/perfil.component';
     Grafica1Component,
     PagesComponent,
     AccountSettingsComponent,
-    PerfilComponent,
   ],
   exports: [
     DashboardComponent,
@@ -38,8 +36,7 @@ import { PerfilComponent } from './perfil/perfil.component';
     FormsModule,
     SharedModule,
     RouterModule,
-    ComponentsModule,
-    ReactiveFormsModule,
+    ComponentsModule
   ]
 })
 export class PagesModule { }
