+++ src/app/pages/pages.routing.ts
@@ -6,20 +6,16 @@ import { DashboardComponent } from './dashboard/dashboard.component';
 import { ProgressComponent } from './progress/progress.component';
 import { Grafica1Component } from './grafica1/grafica1.component';
 import { AccountSettingsComponent } from './account-settings/account-settings.component';
-import { AuthGuard } from '../guards/auth.guard';
-import { PerfilComponent } from './perfil/perfil.component';
 
 const routes: Routes = [
     {
         path: 'dashboard',
         component: PagesComponent,
-        canActivate: [ AuthGuard ],
         children: [
             { path: '', component: DashboardComponent },
             { path: 'progress', component: ProgressComponent },
             { path: 'grafica1', component: Grafica1Component },
             { path: 'account-settings', component: AccountSettingsComponent },
-            { path: 'perfil', component: PerfilComponent, data: { titulo: 'Perfil' } },
         ]
     },
 ];
