--- src/app/pages/perfil/perfil.component.html
@@ -1,58 +1,0 @@
-<div class="row">
-
-  <div class="col-lg-8">
-    <div class="card">
-        <div class="card-body">
-            <h4 class="card-title">Mi perfil de usuario</h4>
-            <h6 class="card-subtitle">Ajustes</h6>
-            <form class="form p-t-20" [formGroup]="perfilForm"  (submit)="actualizarPerfil()">
-                <div class="form-group">
-                    <label for="exampleInputuname">Nombre</label>
-                    <div class="input-group">
-                        <div class="input-group-addon"><i class="ti-user"></i></div>
-                        <input type="text" class="form-control" id="exampleInputuname" formControlName="nombre" placeholder="Nombre de Usuario">
-                    </div>
-                </div>
-                <div class="form-group">
-                    <label for="exampleInputEmail1">Email</label>
-                    <div class="input-group">
-                        <div class="input-group-addon"><i class="ti-email"></i></div>
-                        <input
-                              type="email"
-                              class="form-control"
-                              id="exampleInputEmail1"
-                              formControlName="email"
-                              placeholder="Email"
-                              [readOnly]="usuario.google"
-                        >
-                    </div>
-                </div>
-                <button type="submit"
-                        class="btn btn-success btn-block waves-effect waves-light m-r-10"
-                        [disabled]="perfilForm.invalid">
-                  <i class="fa fa-save" > </i>
-                  Aceptar
-                </button>
-            </form>
-        </div>
-    </div>
-</div>
-<div class="col-lg-4">
-  <div class="card">
-    <div class="card-body">
-      <h4>Avatar</h4>
-      <img [src]="usuario.imagenUrl" class="img-avatar" >
-      <br>
-      <br>
-      <input type="file" (change)="cambiarImagen($any($event.target).files[0])">
-      <br>
-      <br>
-      <button type="submit" [disabled]="!imagenSubir" (click)="subirImagen()" class="btn btn-warning btn-block waves-effect waves-light m-r-10">
-        <i class="fa fa-save" > </i>
-        Guardar Imagen
-      </button>
-    </div>
-  </div>
-</div>
-
-</div>
