--- src/app/services/sidebar.service.ts
@@ -1,21 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class SidebarService {
-
-  menu: any[] = [
-    {
-      titulo: 'Principal',
-      icono: 'mdi mdi-gauge',
-      submenu: [
-        {titulo: 'Main', url: '/'},
-        {titulo: 'Progressbar', url: '/dashboard/progress'},
-        {titulo: 'Graficas', url: '/dashboard/grafica1'},
-      ]
-    }
-  ];
-
-  constructor() { }
-}
