+++ src/app/shared/header/header.component.html
@@ -19,7 +19,7 @@
                 <!-- Logo text --><span>
                     <!-- dark Logo text -->
                     <img src="./assets/images/logo-text.png" alt="homepage" class="dark-logo" />
-                    <!-- Light Logo text -->
+                    <!-- Light Logo text -->    
                     <img src="./assets/images/logo-light-text.png" class="light-logo" alt="homepage" /></span> </a>
         </div>
         <!-- ============================================================== -->
@@ -156,32 +156,25 @@
                 <!-- Profile -->
                 <!-- ============================================================== -->
                 <li class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle waves-effect waves-dark" href="" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                      <img [src]="usuario.imagenUrl" alt="user" class="profile-pic" /></a>
+                    <a class="nav-link dropdown-toggle waves-effect waves-dark" href="" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"><img src="./assets/images/users/1.jpg" alt="user" class="profile-pic" /></a>
                     <div class="dropdown-menu dropdown-menu-right animated fadeIn fast">
                         <ul class="dropdown-user">
                             <li>
                                 <div class="dw-user-box">
-                                    <div class="u-img"><img [src]="usuario.imagenUrl" alt="user"></div>
+                                    <div class="u-img"><img src="./assets/images/users/1.jpg" alt="user"></div>
                                     <div class="u-text">
-                                        <h4>
-                                          {{ usuario.nombre | slice:0:15 }}{{ (usuario.nombre.length > 20) ? '...' : '' }}
-                                        </h4>
-                                        <p class="text-muted">
-                                          {{ usuario.email | slice:0:20 }}{{ (usuario.email.length > 20) ? '...' : '' }}
-                                        </p>
-
-                                        <a href="pages-profile.html" class="btn btn-rounded btn-danger btn-sm">View Profile</a></div>
+                                        <h4>Steave Jobs</h4>
+                                        <p class="text-muted">varun@gmail.com</p><a href="pages-profile.html" class="btn btn-rounded btn-danger btn-sm">View Profile</a></div>
                                 </div>
                             </li>
                             <li role="separator" class="divider"></li>
-                            <li><a routerLink="./perfil"><i class="ti-user"></i> My Profile</a></li>
+                            <li><a href="#"><i class="ti-user"></i> My Profile</a></li>
                             <li><a href="#"><i class="ti-wallet"></i> My Balance</a></li>
                             <li><a href="#"><i class="ti-email"></i> Inbox</a></li>
                             <li role="separator" class="divider"></li>
-                            <li><a routerLink="/dashboard/account-settings"><i class="ti-settings"></i> Account Setting</a></li>
+                            <li><a href="#"><i class="ti-settings"></i> Account Setting</a></li>
                             <li role="separator" class="divider"></li>
-                            <li><a (click)="logout()"><i class="fa fa-power-off"></i> Logout</a></li>
+                            <li><a href="#"><i class="fa fa-power-off"></i> Logout</a></li>
                         </ul>
                     </div>
                 </li>
