+++ src/app/shared/sidebar/sidebar.component.html
@@ -8,30 +8,26 @@
         <nav class="sidebar-nav">
             <ul id="sidebarnav">
                 <li class="user-profile">
-                    <a class="has-arrow waves-effect waves-dark" href="#" aria-expanded="false"><img [src]="usuario.imagenUrl" alt="user" />
-                      <span class="hide-menu">{{ usuario.nombre }}</span></a>
+                    <a class="has-arrow waves-effect waves-dark" href="#" aria-expanded="false"><img src="./assets/images/users/profile.png" alt="user" /><span class="hide-menu">Steave Jobs </span></a>
                     <ul aria-expanded="false" class="collapse">
-                        <li><a routerLink="./perfil">Perfil</a></li>
+                        <li><a href="javascript:void()">My Profile </a></li>
                         <li><a href="javascript:void()">My Balance</a></li>
                         <li><a href="javascript:void()">Inbox</a></li>
-                        <li><a routerLink="/dashboard/account-settings">Account Setting</a></li>
-                        <li><a routerLink="/login">Logout</a></li>
+                        <li><a href="javascript:void()">Account Setting</a></li>
+                        <li><a href="javascript:void()">Logout</a></li>
                     </ul>
                 </li>
                 <li class="nav-devider"></li>
                 <li class="nav-small-cap">PERSONAL</li>
-                <li *ngFor="let item of menuItem">
-                  <a class="has-arrow waves-effect waves-dark" href="#" aria-expanded="false">
-                    <i class="mdi mdi-gauge"></i>
-                      <span class="hide-menu">{{ item.titulo }}
-                        <span class="label label-rouded label-themecolor pull-right">{{ item.submenu.length }}</span>
-                      </span>
-                  </a>
+                <li> <a class="has-arrow waves-effect waves-dark" href="#" aria-expanded="false"><i class="mdi mdi-gauge"></i><span class="hide-menu">Dashboard <span class="label label-rouded label-themecolor pull-right">4</span></span></a>
                     <ul aria-expanded="false" class="collapse">
-                        <!-- <li *ngFor="let sub of item.submenu" ><a routerLink="{{ sub.url }}">{{ sub.titulo }}</a></li> -->
-                        <li *ngFor="let sub of item.submenu" ><a [routerLink]="sub.url">{{ sub.titulo }}</a></li>
+                        <li><a href="index.html">Minimal </a></li>
+                        <li><a href="index2.html">Analytical</a></li>
+                        <li><a href="index3.html">Demographical</a></li>
+                        <li><a href="index4.html">Modern</a></li>
                     </ul>
                 </li>
+                
             </ul>
         </nav>
         <!-- End Sidebar navigation -->
