+++ src/assets/js/custom.js
@@ -4,161 +4,154 @@ Author: Wrappixel
 Email: niravjoshi87@gmail.com
 File: js
 */
+$(function() {
+    "use strict";
+    $(function() {
+        $(".preloader").fadeOut();
+    });
+    jQuery(document).on('click', '.mega-dropdown', function(e) {
+        e.stopPropagation()
+    });
+    // ============================================================== 
+    // This is for the top header part and sidebar part
+    // ==============================================================  
+    var set = function() {
+        var width = (window.innerWidth > 0) ? window.innerWidth : this.screen.width;
+        var topOffset = 0;
+        if (width < 1170) {
+            $("body").addClass("mini-sidebar");
+            $('.navbar-brand span').hide();
+            $(".sidebartoggler i").addClass("ti-menu");
+        } else {
+            $("body").removeClass("mini-sidebar");
+            $('.navbar-brand span').show();
+        }
 
-const customInitFunctions = () => {
+        var height = ((window.innerHeight > 0) ? window.innerHeight : this.screen.height) - 1;
+        height = height - topOffset;
+        if (height < 1) height = 1;
+        if (height > topOffset) {
+            $(".page-wrapper").css("min-height", (height) + "px");
+        }
 
-  $(function() {
-      "use strict";
-      $(function() {
-          $(".preloader").fadeOut();
-      });
-      jQuery(document).on('click', '.mega-dropdown', function(e) {
-          e.stopPropagation()
-      });
-      // ==============================================================
-      // This is for the top header part and sidebar part
-      // ==============================================================
-      var set = function() {
-          var width = (window.innerWidth > 0) ? window.innerWidth : this.screen.width;
-          var topOffset = 0;
-          if (width < 1170) {
-              $("body").addClass("mini-sidebar");
-              $('.navbar-brand span').hide();
-              $(".sidebartoggler i").addClass("ti-menu");
-          } else {
-              $("body").removeClass("mini-sidebar");
-              $('.navbar-brand span').show();
-          }
+    };
+    $(window).ready(set);
+    $(window).on("resize", set);
 
-          var height = ((window.innerHeight > 0) ? window.innerHeight : this.screen.height) - 1;
-          height = height - topOffset;
-          if (height < 1) height = 1;
-          if (height > topOffset) {
-              $(".page-wrapper").css("min-height", (height) + "px");
-          }
+    // ============================================================== 
+    // Theme options
+    // ==============================================================     
+    $(".sidebartoggler").on('click', function() {
+        if ($("body").hasClass("mini-sidebar")) {
+            $("body").trigger("resize");
+            $("body").removeClass("mini-sidebar");
+            $('.navbar-brand span').show();
+            
+        } else {
+            $("body").trigger("resize");
+            $("body").addClass("mini-sidebar");
+            $('.navbar-brand span').hide();
+            
+        }
+    });
 
-      };
-      $(window).ready(set);
-      $(window).on("resize", set);
+    // this is for close icon when navigation open in mobile view
+    $(".nav-toggler").click(function() {
+        $("body").toggleClass("show-sidebar");
+        $(".nav-toggler i").toggleClass("ti-menu");
+        $(".nav-toggler i").addClass("ti-close");
+    });
 
-      // ==============================================================
-      // Theme options
-      // ==============================================================
-      $(".sidebartoggler").on('click', function() {
-          if ($("body").hasClass("mini-sidebar")) {
-              $("body").trigger("resize");
-              $("body").removeClass("mini-sidebar");
-              $('.navbar-brand span').show();
+    $(".search-box a, .search-box .app-search .srh-btn").on('click', function() {
+        $(".app-search").toggle(200);
+    });
+    // ============================================================== 
+    // Right sidebar options
+    // ============================================================== 
+    $(".right-side-toggle").click(function() {
+        $(".right-sidebar").slideDown(50);
+        $(".right-sidebar").toggleClass("shw-rside");
+    });
+    // ============================================================== 
+    // This is for the floating labels
+    // ============================================================== 
+    $('.floating-labels .form-control').on('focus blur', function(e) {
+        $(this).parents('.form-group').toggleClass('focused', (e.type === 'focus' || this.value.length > 0));
+    }).trigger('blur');
 
-          } else {
-              $("body").trigger("resize");
-              $("body").addClass("mini-sidebar");
-              $('.navbar-brand span').hide();
+    // ============================================================== 
+    // Auto select left navbar
+    // ============================================================== 
+    $(function() {
+        var url = window.location;
+        var element = $('ul#sidebarnav a').filter(function() {
+            return this.href == url;
+        }).addClass('active').parent().addClass('active');
+        while (true) {
+            if (element.is('li')) {
+                element = element.parent().addClass('in').parent().addClass('active');
+            } else {
+                break;
+            }
+        }
 
-          }
-      });
+    });
+    // ============================================================== 
+    //tooltip
+    // ============================================================== 
+    $(function() {
+        $('[data-toggle="tooltip"]').tooltip()
+    })
+    // ============================================================== 
+    //Popover
+    // ============================================================== 
+    $(function() {
+        $('[data-toggle="popover"]').popover()
+    })
+    // ============================================================== 
+    // Sidebarmenu
+    // ============================================================== 
+    $(function() {
+        $('#sidebarnav').AdminMenu();
+    });
 
-      // this is for close icon when navigation open in mobile view
-      $(".nav-toggler").click(function() {
-          $("body").toggleClass("show-sidebar");
-          $(".nav-toggler i").toggleClass("ti-menu");
-          $(".nav-toggler i").addClass("ti-close");
-      });
+    // ============================================================== 
+    // Perfact scrollbar
+    // ============================================================== 
+    $('.scroll-sidebar, .right-side-panel, .message-center, .right-sidebar').perfectScrollbar();
+    
+    // ============================================================== 
+    // Resize all elements
+    // ============================================================== 
+    $("body").trigger("resize");
+    // ============================================================== 
+    // To do list
+    // ============================================================== 
+    $(".list-task li label").click(function() {
+        $(this).toggleClass("task-done");
+    });
 
-      $(".search-box a, .search-box .app-search .srh-btn").on('click', function() {
-          $(".app-search").toggle(200);
-      });
-      // ==============================================================
-      // Right sidebar options
-      // ==============================================================
-      $(".right-side-toggle").click(function() {
-          $(".right-sidebar").slideDown(50);
-          $(".right-sidebar").toggleClass("shw-rside");
-      });
-      // ==============================================================
-      // This is for the floating labels
-      // ==============================================================
-      $('.floating-labels .form-control').on('focus blur', function(e) {
-          $(this).parents('.form-group').toggleClass('focused', (e.type === 'focus' || this.value.length > 0));
-      }).trigger('blur');
+    
 
-      // ==============================================================
-      // Auto select left navbar
-      // ==============================================================
-      $(function() {
-          var url = window.location;
-          var element = $('ul#sidebarnav a').filter(function() {
-              return this.href == url;
-          }).addClass('active').parent().addClass('active');
-          while (true) {
-              if (element.is('li')) {
-                  element = element.parent().addClass('in').parent().addClass('active');
-              } else {
-                  break;
-              }
-          }
+    // ============================================================== 
+    // Collapsable cards
+    // ==============================================================
+    $('a[data-action="collapse"]').on('click', function(e) {
+        e.preventDefault();
+        $(this).closest('.card').find('[data-action="collapse"] i').toggleClass('ti-minus ti-plus');
+        $(this).closest('.card').children('.card-body').collapse('toggle');
 
-      });
-      // ==============================================================
-      //tooltip
-      // ==============================================================
-      $(function() {
-          $('[data-toggle="tooltip"]').tooltip()
-      })
-      // ==============================================================
-      //Popover
-      // ==============================================================
-      $(function() {
-          $('[data-toggle="popover"]').popover()
-      })
-      // ==============================================================
-      // Sidebarmenu
-      // ==============================================================
-      $(function() {
-          $('#sidebarnav').AdminMenu();
-      });
+    });
+    // Toggle fullscreen
+    $('a[data-action="expand"]').on('click', function(e) {
+        e.preventDefault();
+        $(this).closest('.card').find('[data-action="expand"] i').toggleClass('mdi-arrow-expand mdi-arrow-compress');
+        $(this).closest('.card').toggleClass('card-fullscreen');
+    });
 
-      // ==============================================================
-      // Perfact scrollbar
-      // ==============================================================
-      $('.scroll-sidebar, .right-side-panel, .message-center, .right-sidebar').perfectScrollbar();
+    // Close Card
+    $('a[data-action="close"]').on('click', function() {
+        $(this).closest('.card').removeClass().slideUp('fast');
+    });
 
-      // ==============================================================
-      // Resize all elements
-      // ==============================================================
-      $("body").trigger("resize");
-      // ==============================================================
-      // To do list
-      // ==============================================================
-      $(".list-task li label").click(function() {
-          $(this).toggleClass("task-done");
-      });
-
-
-
-      // ==============================================================
-      // Collapsable cards
-      // ==============================================================
-      $('a[data-action="collapse"]').on('click', function(e) {
-          e.preventDefault();
-          $(this).closest('.card').find('[data-action="collapse"] i').toggleClass('ti-minus ti-plus');
-          $(this).closest('.card').children('.card-body').collapse('toggle');
-
-      });
-      // Toggle fullscreen
-      $('a[data-action="expand"]').on('click', function(e) {
-          e.preventDefault();
-          $(this).closest('.card').find('[data-action="expand"] i').toggleClass('mdi-arrow-expand mdi-arrow-compress');
-          $(this).closest('.card').toggleClass('card-fullscreen');
-      });
-
-      // Close Card
-      $('a[data-action="close"]').on('click', function() {
-          $(this).closest('.card').removeClass().slideUp('fast');
-      });
-
-  });
-
-}
-
-customInitFunctions();
+});
