+++ src/environments/environment.prod.ts
@@ -1,6 +1,3 @@
-//TODO: Cambiar el url de los servicios
-
 export const environment = {
-  production: true,
-  base_url: 'http://localhost:3005/api'
+  production: true
 };
