+++ src/index.html
@@ -5,9 +5,6 @@
   <title>AdminPro</title>
   <base href="/">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-
-  <script src="https://accounts.google.com/gsi/client" async defer></script>
-
   <link rel="icon" type="image/x-icon" href="favicon.ico">
 
   <!-- Favicon icon -->
@@ -59,7 +56,7 @@
   <script src="./assets/plugins/sticky-kit-master/dist/sticky-kit.min.js"></script>
   <script src="./assets/plugins/sparkline/jquery.sparkline.min.js"></script>
   <!--Custom JavaScript -->
-  <script src="./assets/js/custom.js"></script>
+  <script src="./assets/js/custom.min.js"></script>
 
   <script src="https://cdn.jsdelivr.net/npm/chart.js@2.8.0"></script>
