+++ src/styles.css
@@ -9,12 +9,3 @@
 #themecolors .selector{
   cursor: pointer;
 }
-
-.img-avatar{
-  width: 200px;
-  height: 200px;
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  border-radius: 25%
-}
